feat: move render file to locations folder
